indexer: move MemorySyncCursorStore to test-only compilation unit
Move `MemorySyncCursorStore` out of production
`indexer/sync_client.go` into a dedicated `_test.go` file.

What changed:
- Removed `MemorySyncCursorStore` and constructor from
  `indexer/sync_client.go`.
- Added identical implementation to
  `indexer/memory_sync_cursor_store_test.go`.
- Kept the same public test helper API
  (`NewMemorySyncCursorStore`) for existing tests, so no
  test-callsite rewrites were needed.

Why this is a low-risk cleanup:
- Repo-wide usage check shows the memory cursor store is only
  referenced by tests inside the `indexer` package.
- Keeping test scaffolding in production code expanded package
  surface without runtime value.
- Moving it to a test-only file removes dead runtime code and
  keeps the production package focused on real sync logic.

Behavioral impact:
- No runtime behavior changes.
- No wire/proto, persistence, or FSM semantics affected.
- Test behavior stays the same because the helper remains
  available during test builds.

Validation:
- go test ./indexer/...
- go test ./oor/... ./serverconn/... ./indexer/...
